fix: updated custom-tree-layout
