+++ projects/observability/src/shared/components/topology/d3/layouts/custom-tree-layout.ts
@@ -1,56 +1,59 @@
-import { hierarchy, HierarchyNode } from 'd3-hierarchy';
-import { RenderableTopology, TopologyEdge, TopologyNode } from '../../topology';
-import { D3ProxyNode, TreeLayout } from './tree-layout';
 import cytoscape from 'cytoscape';
+import { cloneDeep } from 'lodash-es';
+import { RenderableTopology, TopologyEdge, TopologyNode } from '../../topology';
+import { TreeLayout } from './tree-layout';
 
 export class CustomTreeLayout extends TreeLayout {
   public layout(topology: RenderableTopology<TopologyNode, TopologyEdge>): void {
+    const selfNode = topology.nodes.find(node => node.outgoing.length > 0);
+    const selfLink = cloneDeep(selfNode?.outgoing[0]);
+
+    if (selfLink && selfNode) {
+      console.log('selfNode: ', selfNode);
+
+      selfLink.source = selfNode;
+      selfLink.target = topology.nodes[2];
+
+      selfNode.outgoing.push(selfLink);
+      topology.nodes[2].incoming.push(selfLink);
+      topology.edges.push(selfLink);
+    }
+
     console.log(topology);
-    // console.log(topology.nodes.map(node => node.userNode.data?.name));
-    // var g = new dracula.Graph();
-    const rootHierarchyNode = hierarchy(this.buildHierarchyProxyNodes(topology.nodes, { x: 0, y: 0 }));
+
+    const edges = topology.edges.map(edge => ({
+      data: { source: edge.source.userNode.data?.name!, target: edge.target.userNode.data?.name! }
+    }));
 
     const rootHierarchyNode2 = cytoscape({
       elements: {
         nodes: topology.nodes.map(node => ({ data: { id: node.userNode.data?.name! } })),
-        edges: topology.edges.map(edge => ({
-          data: { source: edge.source.userNode.data?.name!, target: edge.target.userNode.data?.name! }
-        }))
+        edges: edges
       },
       layout: {
         name: 'breadthfirst',
-        avoidOverlap: true
+        spacingFactor: 300
       }
     });
 
-    console.log(rootHierarchyNode);
-    console.log(rootHierarchyNode2.json());
+    const graph: Elements = rootHierarchyNode2.json() as Elements;
 
-    const nodeWidth = this.getNodeWidth(rootHierarchyNode);
-    const nodeHeight = this.getNodeHeight(rootHierarchyNode);
+    const map: Map<string, { x: number; y: number }> = new Map<string, { x: number; y: number }>(
+      graph.elements.nodes.map(node => [node.data.id, { x: node.position.y + 455, y: node.position.x / 4 + 470 }])
+    );
 
-    this.updateLayout(rootHierarchyNode, 0, -1, nodeWidth * 1.2, nodeHeight * 1.2);
-
-    // console.log(topology);
-    console.log(rootHierarchyNode);
-  }
-
-  public updateLayout(
-    hierarchyNode: HierarchyNode<D3ProxyNode>,
-    nodeRowIndex: number,
-    nodeColumnIndex: number,
-    cellWidth: number,
-    cellHeight: number
-  ): number {
-    if (hierarchyNode.data.sourceNode !== undefined) {
-      hierarchyNode.data.sourceNode.x = nodeColumnIndex * cellWidth;
-      hierarchyNode.data.sourceNode.y = nodeRowIndex * cellHeight;
-    }
-
-    hierarchyNode.children?.forEach((node, index) => {
-      this.updateLayout(node, nodeRowIndex + index, nodeColumnIndex + 1, cellWidth, cellHeight);
+    topology.nodes.forEach(node => {
+      node.x = map.get(node.userNode.data?.name!)!.x;
+      node.y = map.get(node.userNode.data?.name!)!.y;
     });
-
-    return hierarchyNode.children?.length ?? 1;
   }
+}
+
+interface Elements {
+  elements: {
+    nodes: {
+      data: { id: string };
+      position: { x: number; y: number };
+    }[];
+  };
 }
